add single-precision (WP=4) build support for standalone fesom.x
Make working precision configurable via -DUSE_SINGLE_PRECISION=ON. Default
build stays double precision and unchanged.

- WP is now preprocessor-selected (4/8) in o_PARAM; central MPI_WP parameter
  in MOD_PARTIT mirrors it (MPI_REAL vs MPI_DOUBLE_PRECISION).
- CMake gains USE_SINGLE_PRECISION option with per-compiler default-real flag
  flip (Intel -r4, GNU drops -fdefault-real-8, Cray -s real32, NVHPC -Mr4);
  CVMix forced OFF in SP (double-only, out of scope).
- Replace MPI_DOUBLE_PRECISION/MPI_DOUBLE with MPI_WP at WP-data sites; convert
  WP comm/gather buffers from real(real64) to real(kind=WP).
- NetCDF stays double on disk; WP reads routed through generic nf_*_vara_x so
  existing meshes/restarts/forcing keep working (convert-on-I/O).
- mod_transit and reference-density locals converted to WP; io_meandata vector
  rotation round-trips through WP temps; iceberg subsystem made WP-consistent.
- Startup banner prints WP byte size and SINGLE/DOUBLE mode.

+++ src/icb_modules.F90
@@ -116,7 +116,7 @@ save
   real,dimension(:), allocatable:: arr_block
   integer,dimension(:), allocatable:: elem_block
   integer,dimension(:), allocatable:: pe_block
-  real(real64), dimension(:), allocatable:: elem_area_glob
+  real(kind=WP), dimension(:), allocatable:: elem_area_glob
   real,dimension(:), allocatable:: vl_block
 
   !array for output in netcdf
